chore: refactor web api endpoints
